+++ R/get.R
@@ -6,6 +6,23 @@
 #' @return the class of the object returned depends on the function, and on their arguments. Typically, a data.frame or a vector if the output can be reduced to one line.
 #'
 #' @description Helpful functions to get content from a `mrgmod` object (i.e. data) or from a `mapbayests` object (`data`, `eta`, `cov`, `param`, `phi`).
+#'
+#' @examples
+#' # From a model object (mrgmod)
+#' mod <- exmodel(compile = FALSE)
+#' get_data(mod)
+#'
+#' # From an estimation object (mapbayests)
+#' get_data(est001)
+#' get_data(est001, output = "list")
+#'
+#' get_eta(est001)
+#' get_eta(est001, output = "list")
+#'
+#' get_cov(est001)
+#'
+#' get_phi(est001)
+#'
 NULL
 #> NULL
 
@@ -18,8 +35,8 @@ get_data <- function(x, ...)UseMethod("get_data")
 #' @param x model object
 #' @param ... not used
 #'
+#' @rdname get_x
 #' @method get_data mrgmod
-#' @return a tibble
 #' @export
 get_data.mrgmod <- function(x, ...){
   as_tibble(x@args$data)
@@ -31,8 +48,8 @@ get_data.mrgmod <- function(x, ...){
 #' @param ... not used
 #' @param output either a single data.frame ("df", the default) or a list ("list") of individual data sets
 #'
+#' @rdname get_x
 #' @method get_data mapbayests
-#' @return a tibble
 #' @export
 get_data.mapbayests <- function(x, ..., output = "df"){
   iddata <- map(x$arg.ofv.id, ~ devalid_data_set(.x$idvaliddata))
@@ -52,8 +69,8 @@ get_eta <- function(x, ...) UseMethod("get_eta")
 #' @param ... not used
 #' @param output either a list ("list"), a data.frame ("df") or a vector of numeric ("num"). Default to "num" if only one ID.
 #'
+#' @rdname get_x
 #' @method get_eta mapbayests
-#' @return a tibble
 #' @export
 get_eta.mapbayests <- function(x, ..., output = NULL){
 
@@ -105,8 +122,8 @@ get_cov <- function(x, ...) UseMethod("get_cov")
 #' @param ... not used
 #' @param simplify a logical. If TRUE (the default) and only one ID, one matrix is returned instead of a list of length 1
 #'
+#' @rdname get_x
 #' @method get_cov mapbayests
-#' @return a tibble
 #' @export
 get_cov.mapbayests <- function(x, ..., simplify = TRUE){
   ans <- x$covariance
@@ -126,8 +143,8 @@ get_param <- function(x, ...) UseMethod("get_param")
 #' @param keep_ID a logical. By default, the ID variable is dropped if one ID in data.
 #' @param keep_names a logical. By default, names are dropped if one parameter is requested, and output is not a data frame.
 #'
+#' @rdname get_x
 #' @method get_param mapbayests
-#' @return a tibble
 #' @export
 get_param.mapbayests <- function(x, ..., output = NULL, keep_ID = NULL, keep_names = NULL){
   #Check Arguments
@@ -208,10 +225,10 @@ get_phi <- function(x, ...) UseMethod("get_phi")
 
 #' Return "NONMEM phi"-like file from a mapbayests
 #'
+#' @rdname get_x
 #' @param x mapbayests object
 #' @param ... not used
 #' @method get_phi mapbayests
-#' @return a tibble
 #' @export
 get_phi.mapbayests <- function(x, ...){
   namcov <- namephicov(n_eta(x$model))