+++ inst/exmodel/mrg_301.cpp
@@ -1,0 +1,44 @@
+$PROB Reference model 
+
+$PARAM @annotated
+TVCL   : 4.00 : Clearance (L/h)
+TVVC   : 70.0 : Central volume of distribution (L)
+TVKA   : 1.00 : Absorption rate (h-1)
+
+ETA1 : 0 : CL
+ETA2 : 0 : VC
+ETA3 : 0 : KA
+
+$PARAM @annotated @covariates
+BW : 75 : Body weight (kg)
+SEX:  0 : Sex (0 Male, 1 Female)
+
+$OMEGA
+0.2 // CL
+0.2 // VC
+0.2 // KA 
+
+$SIGMA 
+0.05 // err prop
+0   //  err additive 
+
+
+$CMT @annotated
+DEPOT   : Depot () [ADM]
+CENTRAL : Central () [OBS]
+
+$TABLE
+double DV = (CENTRAL / VC) * (1 + EPS(1)) + EPS(2) ;
+
+$MAIN
+double CL  = TVCL  * exp(ETA(1) + ETA1 ) * pow(BW/75, 1.5) * pow(.75, SEX) ; 
+double VC  = TVVC  * exp(ETA(2) + ETA2 ) ;
+double KA  = TVKA  * exp(ETA(3) + ETA3 ) ;
+
+double K20 = CL / VC ;
+
+$ODE
+dxdt_DEPOT   = - KA * DEPOT ;
+dxdt_CENTRAL = - K20 * CENTRAL + KA * DEPOT ;
+
+$CAPTURE DV