+++ tests/testthat/test-read_mrgsolve_model.R
@@ -1,7 +1,26 @@
 test_that("read cmts", {
-  mod1 <- mread("ex_mbr1", mbrlib())
-  mod2 <- mread("ex_mbr2", mbrlib())
-  mod3 <- mread("ex_mbr3", mbrlib())
+  mod1 <- mcode("mod1", "
+$PARAM DUR = 1
+$CMT @annotated
+DEPOT : Depot compartment () [ADM]
+CENT : Central compartment ()[ADM, OBS]
+", compile = FALSE)
+
+
+  mod2 <- mcode("mod1", "
+$PARAM DUR = 1
+$CMT @annotated
+DEPOT : Depot compartment () [ADM]
+CENT : Central compartment ()[OBS]
+PERIPH: Central compartment ()
+METAB: Central compartment ()[OBS]
+", compile = FALSE)
+
+  mod3 <- mcode("mod1", "
+$CMT @annotated
+CENT : Central compartment ()[ADM, OBS]
+", compile = FALSE)
+
   expect_equal(adm_cmt(mod1), c(1,2))
   expect_equal(adm_cmt(mod2), 1)
   expect_equal(adm_cmt(mod3), 1)
@@ -22,8 +41,6 @@ test_that("Detection of cmt to fit in data is correct", {
   ##3) If more than one compartment with observation in the dataset (in one individual) : stop, user should use [OBS] if multiple obs compartments are used
   ##4) Compartment must exist in the model
   ##5) Sigma must be equal to 2
-  ##6) mapbayests methods must work too
-
 
   basecode <- "
 $PARAM @annotated
@@ -59,7 +76,7 @@ CENT  : Central compartment (mg/L)[ADM, OBS]
 PERIPH: Peripheral compartment ()
   "
 
-  mod1 <- mcode("mod1", code = paste0(basecode, sigma0, cmt0))
+  mod1 <- mcode("mod1", code = paste0(basecode, sigma0, cmt0), compile = FALSE)
   data1 <- mod1 %>%
     adm_lines(amt = 100) %>%
     obs_lines(time = 24, DV = 6) %>%
@@ -67,7 +84,7 @@ PERIPH: Peripheral compartment ()
     get_data()
 
   #This works
-  expect_s3_class(mapbayest(mod1, data1, verbose = F), "mapbayests")
+  expect_error(check_mapbayr_modeldata(mod1, data1), NA)
 
   #1 Modified code to test point 1 = the length of sigma matrix diagonal must be 2 x number of compartments
   sigma1 <- "$SIGMA
@@ -75,53 +92,49 @@ PERIPH: Peripheral compartment ()
 0.1 // additive
 0.2
   "
-  mcode("mod1", code = paste0(basecode, sigma1, cmt0)) %>%
-    mapbayest(data = data1) %>%
-    expect_error(".*SIGMA: Define one pair of sigma values.*")
+  merge_errors <- function(x){paste(x$descr, collapse = " ")}
 
-  #2 Modified data to test point 2 = obs in data must be defined in model
+  mod1 <- mcode("mod1", code = paste0(basecode, sigma1, cmt0), compile = FALSE)
+
+  expect_match(merge_errors(check_mapbayr_model(mod1, check_compile = FALSE)),
+               "\\$SIGMA: Define one pair of sigma values.*")
+
+  #2 Modified data to test point 2 = obs cmt in data must be defined in model
   data2 <- mutate(data1, cmt = c(1, 1, 99))
-  expect_error(mapbayest(x = mod1, data2), "ID =1; CMT =99\n One or more compartment with observation .* in data don't match those defined with")
+  expect_error(check_mapbayr_modeldata(x = mod1, data2), "ID =1; CMT =99\n One or more compartment with observation .* in data don't match those defined with")
 
   #Other type of model. Here [OBS] is NOT defined in $CMT
   cmt3 <- "$CMT CENT PERIPH"
-  mod3 <- mcode("mod3", code = paste0(basecode, sigma0, cmt3))
+  mod3 <- mcode("mod3", code = paste0(basecode, sigma0, cmt3), compile = FALSE)
 
-  check3 <- check_mapbayr_model(mod3)
+  check3 <- check_mapbayr_model(mod3, check_compile = FALSE)
   expect_true("$CMT: No [ADM] compartment(s) defined (optionnal)." %in% check3$descr)
   expect_true("$CMT: No [OBS] compartment(s) defined (optionnal)." %in% check3$descr)
   expect_equal(fit_cmt(mod3, data1), 1)
   expect_null(obs_cmt(mod3))
 
   #This works
-  expect_s3_class(est3 <- mapbayest(mod3, data1, verbose = F), "mapbayests")
+  expect_error(check_mapbayr_modeldata(mod3, data1), NA)
 
   ##3) Only one obs compartment in data
   data3 <- mutate(data1, cmt = c(1,2,1)) #here multiple cmt in one patient (ID 1, cmt 1 2)
-  expect_error(mapbayest(mod3, data3, verbose = F), "More than one 'observation compartment' to detect from data. Consider editing model code")
+  expect_error(check_mapbayr_modeldata(mod3, data3), "More than one 'observation compartment' to detect from data. Consider editing model code")
 
   #also test among two patients (ID 1 cmt 1, ID2 cmt 2) see issue #48
   data31 <- mutate(data1, cmt = c(1,2,2), ID = 20) %>% bind_rows(data1)
-  expect_error(mapbayest(mod3, data31, verbose = F), "More than one 'observation compartment' to detect from data. Consider editing model code")
+  expect_error(check_mapbayr_modeldata(mod3, data31), "More than one 'observation compartment' to detect from data. Consider editing model code")
 
   ##4) Compartment must exist in the model
   data4 <- mutate(data1, cmt = c(1,99,99))
-  mapbayest(mod3, data4) %>% expect_error("Compartment number with observation in dataset does not exist in model.")
+  expect_error(check_mapbayr_modeldata(mod3, data4), "Compartment number with observation in dataset does not exist in model.")
 
   ##5) Sigma must be equal to 2
 
-  mcode("mod5", code = paste0(basecode, sigma1, cmt3)) %>%
-    mapbayest(data = data1) %>%
-    expect_error(".*Define only one pair of sigma values .* if you do not use .*One observation compartment will be defined from MDV=0 lines in individual data")
-
-  ##6) Method
-  expect_output(print(est3))
-  expect_error(as.data.frame(est3), NA)
-  expect_error(augment(est3), NA)
-  expect_error(plot(est3), NA)
-  expect_error(hist(est3), NA)
-
+  mod5 <- mcode("mod5", code = paste0(basecode, sigma1, cmt3), compile = FALSE)
+  check5 <- check_mapbayr_model(mod5, check_compile = FALSE)
+  expect_match(merge_errors(check5), ".*Define only one pair of sigma values .* if you do not use .*One observation compartment will be defined from MDV=0 lines in individual data")
 })
+
 
 
 test_that("wrong code is correctly checked", {
@@ -161,7 +174,7 @@ dxdt_PERIPH =  K12 * CENT - K21 * PERIPH ;
 $CAPTURE DV
 "
 
-  check1 <- check_mapbayr_model(mcode("test1", code = code1))
+  check1 <- check_mapbayr_model(mcode("test1", code = code1, compile = FALSE), check_compile = FALSE)
 
   expect_true("$PARAM: 2 ETA found, but not sequentially named ETA1." %in% check1$descr)
   expect_true("$PARAM: Initial value is not 0 for all ETA." %in% check1$descr)
@@ -208,19 +221,19 @@ double V2 = TVV1 * exp(ETA3 + ETA(3)) ;
 $PKMODEL ncmt = 2, depot = FALSE
 $CAPTURE DV PAR MET CL
 "
-  mod1 <- mcode("mod1", code1)
+  mod1 <- mcode("mod1", code1, compile = FALSE)
 
   data1 <- mod1 %>%
     adm_lines(amt = 100) %>%
     obs_lines(time = 5, DV = 2.2, DVmet = 1.1) %>%
     get_data()
 
-  expect_error(mapbayest(mod1, data1, verbose = FALSE), NA)
-  expect_error(mapbayest(mod1, mutate(data1, BW = 10), verbose = FALSE), NA)
-  expect_error(mapbayest(mod1, mutate(data1, ETA1 = 1), verbose = FALSE), "These variables cannot be set in both model and data: ETA1")
-  expect_error(mapbayest(mod1, mutate(data1, TVCL = 10), verbose = FALSE), "These variables cannot be set in both model and data: TVCL")
-  expect_error(mapbayest(mod1, mutate(data1, CL = 10), verbose = FALSE), "These variables cannot be set in both model and data: CL")
-  expect_error(mapbayest(mod1, mutate(data1, V1 = 99), verbose = FALSE), "These variables cannot be set in both model and data: V1")
+  expect_error(check_mapbayr_modeldata(mod1, data1), NA)
+  expect_error(check_mapbayr_modeldata(mod1, mutate(data1, BW = 10)), NA)
+  expect_error(check_mapbayr_modeldata(mod1, mutate(data1, ETA1 = 1)), "These variables cannot be set in both model and data: ETA1")
+  expect_error(check_mapbayr_modeldata(mod1, mutate(data1, TVCL = 10)), "These variables cannot be set in both model and data: TVCL")
+  expect_error(check_mapbayr_modeldata(mod1, mutate(data1, CL = 10)), "These variables cannot be set in both model and data: CL")
+  expect_error(check_mapbayr_modeldata(mod1, mutate(data1, V1 = 99)), "These variables cannot be set in both model and data: V1")
 
 })
 
@@ -264,10 +277,10 @@ $PKMODEL ncmt = 1, depot = FALSE
   mod3 <- mcode("mod3", code3, compile = FALSE)
   mod4 <- mcode("mod4", code4, compile = FALSE)
 
-  descr1 <- check_mapbayr_model(mod1)$descr
-  descr2 <- check_mapbayr_model(mod2)$descr
-  descr3 <- check_mapbayr_model(mod3)$descr
-  descr4 <- check_mapbayr_model(mod4)$descr
+  descr1 <- check_mapbayr_model(mod1, check_compile = FALSE)$descr
+  descr2 <- check_mapbayr_model(mod2, check_compile = FALSE)$descr
+  descr3 <- check_mapbayr_model(mod3, check_compile = FALSE)$descr
+  descr4 <- check_mapbayr_model(mod4, check_compile = FALSE)$descr
 
   expect_true(!"$SIGMA: Exponential error found. Sigma values in position 2,4... cannot be equal to 0." %in% descr1)
   expect_true(!"$SIGMA: Exponential error found. Sigma values in position 1,3... must be equal to 0." %in% descr1)
@@ -299,10 +312,8 @@ $PKMODEL ncmt = 1, depot = TRUE
 $CAPTURE DV CL
 "
 
-  mod1 <- mcode('mod1', code1)
-  data1 <- data.frame(ID = 1, TIME = c(0, 1.1, 5.2, 12.3), EVID = c(1,0,0,0), AMT = c(500, 0,0,0), CMT = c(1,2,2,2), DV = c(0, 15.1, 29.5, 22.3))
-
-  expect_error(mapbayest(mod1, data1), "Cannot be equal to zero")
+  mod1 <- mcode('mod1', code1, compile = FALSE)
+  expect_match(paste(check_mapbayr_model(mod1, check_compile = FALSE)$descr, collapse = " "), "Cannot be equal to zero")
 
   code2 <- "$PARAM ETA1 = 0, ETA2 = 0,
 KA = 0.5, TVCL = 1.1, TVV = 23.3
@@ -317,8 +328,6 @@ double DV=CENT/V*(1+EPS(1))+EPS(2);
 $PKMODEL ncmt = 1, depot = TRUE
 $CAPTURE DV CL
 "
-  mod2 <- mcode('mod2', code2)
-  expect_error(mapbayest(mod2, data1), "SIGMA are equal to zero")
-
-
+  mod2 <- mcode('mod2', code2, compile = FALSE)
+  expect_match(paste(check_mapbayr_model(mod2, check_compile = FALSE)$descr, collapse = " "), "SIGMA are equal to zero")
 })