+++ tests/testthat/test-verbose.R
@@ -1,10 +1,8 @@
-mod <- mread('ex_mbr1', mbrlib())
-data <- mod %>%
-    adm_lines(amt = 10, addl = 2, ii = 12) %>%
-    obs_lines(DV = c(5, 10), time = c(18, 40)) %>%
-    get_data()
+mod <- exmodel(add_exdata = FALSE)
+dat <- exdata(ID = 2) %>%
+  mutate(DV = c(NA, 200))
 
 test_that("verbose works on difficulty warning", {
-  expect_message(mapbayest(mod, data), "Reset with ")
-  expect_message(mapbayest(mod, data, verbose = FALSE), NA)
+  expect_message(mapbayest(mod, dat), "Reset with ")
+  expect_message(mapbayest(mod, dat, verbose = FALSE), NA)
 })